cleared WP library and a lot of low-hanging fruit for iOS build
