+++ Models/PeakFinder.cs
@@ -1,5 +1,4 @@
-﻿using Android.Net.Eap;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
