+++ Models/WPLibrary.cs
@@ -1,6 +1,4 @@
-﻿using Android.Content.Res;
-using Android.OS;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -329,88 +327,11 @@ namespace EnlightenMAUI.Models
             library[name] = adjusted;
         }
 
-        Java.IO.File traverseDown(string[] rootPath, Java.IO.File dir)
-        {
-            var subs = dir.ListFiles();
-            Java.IO.File libraryFolder = null;
-
-            foreach (var sub in subs)
-            {
-                if (sub.AbsolutePath.Split('/').Last() == rootPath[0])
-                {
-                    if (rootPath.Length == 1)
-                    {
-                        libraryFolder = sub;
-                        break;
-                    }
-
-                    else if (sub.IsDirectory)
-                    {
-                        string[] shortenedPath = new string[rootPath.Length - 1];
-                        Array.Copy(rootPath, 1, shortenedPath, 0, rootPath.Length - 1);
-                        return traverseDown(shortenedPath, sub);
-                    }
-                }
-            }
-
-            return libraryFolder;
-        }
-
         async Task loadFiles(string root, bool doDecon = true, string correctionFileName = "etalon_correction.json")
         {
             isLoading = true;
-            var cacheDirs = Platform.AppContext.GetExternalFilesDirs(null);
-            Java.IO.File libraryFolder = null;
-            string[] rootPath = root.Split('/');
-            int depth = rootPath.Length;
+            await PlatformUtil.loadFiles(root, library, originalRaws, originalDarks, doDecon, correctionFileName);
 
-            foreach (var cDir in cacheDirs)
-            {
-                libraryFolder = traverseDown(rootPath, cDir);
-                if (libraryFolder != null)
-                    break;
-
-            }
-
-            if (libraryFolder == null)
-            {
-                if (library.Count > 0)
-                    loadSucceeded = true;
-                isLoading = false;
-                InvokeLoadFinished();
-                return;
-            }
-
-            Regex csvReg = new Regex(@".*\.csv$");
-            Regex jsonReg = new Regex(@".*\.json$");
-
-            var libraryFiles = libraryFolder.ListFiles();
-
-            foreach (var libraryFile in libraryFiles)
-            {
-                if (jsonReg.IsMatch(libraryFile.AbsolutePath))
-                {
-                    try
-                    {
-                        await loadJSON(libraryFile);
-                    }
-                    catch (Exception e)
-                    {
-                        logger.debug("loading {0} failed with exception {1}", libraryFile.AbsolutePath, e.Message);
-                    }
-                }
-                else if (csvReg.IsMatch(libraryFile.AbsolutePath))
-                {
-                    try
-                    {
-                        await loadCSV(libraryFile);
-                    }
-                    catch (Exception e)
-                    {
-                        logger.debug("loading {0} failed with exception {1}", libraryFile.AbsolutePath, e.Message);
-                    }
-                }
-            }
             if (library.Count > 0)
                 loadSucceeded = true;
             isLoading = false;
@@ -436,226 +357,7 @@ namespace EnlightenMAUI.Models
 
             logger.debug("finished prepping data for decon");
         }
-        async Task loadCSV(string path)
-        {
-            string name = path.Split('/').Last().Split('.').First();
-
-            SimpleCSVParser parser = new SimpleCSVParser();
-            AssetManager assets = Platform.AppContext.Assets;
-            Stream s = assets.Open(path);
-            StreamReader sr = new StreamReader(s);
-            await parser.parseStream(s);
-
-            Measurement m = new Measurement();
-            m.wavenumbers = parser.wavenumbers.ToArray();
-            m.raw = parser.intensities.ToArray();
-            m.excitationNM = 785;
-
-
-#if USE_DECON
-            Deconvolution.Spectrum spec = new Deconvolution.Spectrum(parser.wavenumbers, parser.intensities);
-#endif
-
-            Measurement mOrig = m.copy();
-            originalRaws.Add(name, mOrig.raw);
-
-            /*
-            double[] smoothedSpec = PlatformUtil.ProcessBackground(m.wavenumbers, m.processed);
-            while (smoothedSpec == null || smoothedSpec.Length == 0)
-            {
-                smoothedSpec = PlatformUtil.ProcessBackground(m.wavenumbers, m.processed);
-                await Task.Delay(50);
-            }
-            */
-
-            Measurement updated = wavecal.crossMapWavenumberData(m.wavenumbers, m.raw);
-            double airPLSLambda = 10000;
-            int airPLSMaxIter = 100;
-            double[] array = AirPLS.smooth(updated.processed, airPLSLambda, airPLSMaxIter, 0.001, verbose: false, (int)roiStart, (int)roiEnd);
-            double[] shortened = new double[updated.processed.Length];
-            Array.Copy(array, 0, shortened, roiStart, array.Length);
-            updated.raw = shortened;
-            updated.dark = null;
-
-            library.Add(name, updated);
-
-#if USE_DECON
-            deconvolutionLibrary.library.Add(name, spec);
-#endif
-
-            logger.info("finish loading library file from {0}", path);
-        }
-        async Task loadCSV(Java.IO.File file)
-        {
-            logger.info("start loading library file from {0}", file.AbsolutePath);
-
-            string name = file.AbsolutePath.Split('/').Last().Split('.').First();
-
-            SimpleCSVParser parser = new SimpleCSVParser();
-            Stream s = File.OpenRead(file.AbsolutePath); 
-            StreamReader sr = new StreamReader(s);
-            await parser.parseStream(s);
-
-            Measurement m = new Measurement();
-            m.wavenumbers = parser.wavenumbers.ToArray();
-            m.raw = parser.intensities.ToArray();
-            m.excitationNM = 785;
-
-
-#if USE_DECON
-            Deconvolution.Spectrum spec = new Deconvolution.Spectrum(parser.wavenumbers, parser.intensities);
-#endif
-
-            Measurement mOrig = m.copy();
-            originalRaws.Add(name, mOrig.raw);
-
-            /*
-            double[] smoothedSpec = PlatformUtil.ProcessBackground(m.wavenumbers, m.processed);
-            while (smoothedSpec == null || smoothedSpec.Length == 0)
-            {
-                smoothedSpec = PlatformUtil.ProcessBackground(m.wavenumbers, m.processed);
-                await Task.Delay(50);
-            }
-            */
-
-            if (false)
-            {
-                double[] smoothed = PlatformUtil.ProcessBackground(m.wavenumbers, m.processed, "", 14, 200);
-                double[]  wavenumbers = Enumerable.Range(400, smoothed.Length).Select(x => (double)x).ToArray();
-                Measurement updated = new Measurement();
-                updated.wavenumbers = wavenumbers;
-                updated.raw = smoothed;
-                library.Add(name, updated);
-
-#if USE_DECON
-                Deconvolution.Spectrum upSpec = new Deconvolution.Spectrum(new List<double>(wavenumbers), new List<double>(smoothed));
-                deconvolutionLibrary.library.Add(name, upSpec);
-#endif
-            }
-
-            else
-            {
-                double[] wavenumbers = Enumerable.Range(400, 2008).Select(x => (double)x).ToArray();
-                double[] newIntensities = Wavecal.mapWavenumbers(m.wavenumbers, m.processed, wavenumbers);
-
-                Measurement updated = new Measurement();
-                updated.wavenumbers = wavenumbers;
-                updated.raw = newIntensities;
-                //double airPLSLambda = 10000;
-                //int airPLSMaxIter = 100;
-                //double[] array = AirPLS.smooth(updated.processed, airPLSLambda, airPLSMaxIter, 0.001, verbose: false, (int)roiStart, (int)roiEnd);
-                //double[] shortened = new double[updated.processed.Length];
-                //Array.Copy(array, 0, shortened, roiStart, array.Length);
-                //updated.raw = shortened;
-                //updated.dark = null;
-
-                library.Add(name, updated);
-
-#if USE_DECON
-                Deconvolution.Spectrum upSpec = new Deconvolution.Spectrum(new List<double>(updated.wavenumbers), new List<double>(updated.processed));
-                deconvolutionLibrary.library.Add(name, upSpec);
-#endif
-            }
-
-            logger.info("finish loading library file from {0}", file.AbsolutePath);
-        }
-        async Task loadJSON(string path)
-        {
-            logger.info("start loading library file from {0}", path);
-
-
-
-            string name = path.Split('/').Last().Split('.').First();
-
-            AssetManager assets = Platform.AppContext.Assets;
-            Stream s = assets.Open(path);
-            StreamReader sr = new StreamReader(s);
-            string blob = await sr.ReadToEndAsync();
-
-            spectrumJSON json = JsonConvert.DeserializeObject<spectrumJSON>(blob);
-            if (json.tag != null && json.tag.Length > 0)
-            {
-                name = json.tag;
-            }
-
-            Measurement m = new Measurement(json);
-            Wavecal otherCal = new Wavecal(m.pixels);
-            otherCal.coeffs = m.wavecalCoeffs;
-            otherCal.excitationNM = m.excitationNM;
-
-            Measurement mOrig = m.copy();
-            originalRaws.Add(name, mOrig.raw);
-            originalDarks.Add(name, mOrig.dark);
-
-            Measurement updated = wavecal.crossMapIntensityWavenumber(otherCal, m);
-            double airPLSLambda = 10000;
-            int airPLSMaxIter = 100;
-            double[] array = AirPLS.smooth(updated.processed, airPLSLambda, airPLSMaxIter, 0.001, verbose: false, (int)roiStart, (int)roiEnd);
-            double[] shortened = new double[updated.processed.Length];
-            Array.Copy(array, 0, shortened, roiStart, array.Length);
-            updated.raw = shortened;
-            updated.dark = null;
-
-            library.Add(name, updated);
-            logger.info("finish loading library file from {0}", path);
-        }
-        async Task loadJSON(Java.IO.File file)
-        {
-            logger.info("start loading library file from {0}", file.AbsolutePath);
-
-            string name = file.AbsolutePath.Split('/').Last().Split('.').First();
-
-            SimpleCSVParser parser = new SimpleCSVParser();
-            Stream s = File.OpenRead(file.AbsolutePath);
-            StreamReader sr = new StreamReader(s);
-            string blob = await sr.ReadToEndAsync();
-
-            spectrumJSON json = JsonConvert.DeserializeObject<spectrumJSON>(blob);
-            if (json.tag != null && json.tag.Length > 0)
-            {
-                name = json.tag;
-            }
-
-            Measurement m = new Measurement(json);
-            Wavecal otherCal = new Wavecal(m.pixels);
-            otherCal.coeffs = m.wavecalCoeffs;
-            otherCal.excitationNM = m.excitationNM;
-
-            Measurement mOrig = m.copy();
-            originalRaws.Add(name, mOrig.raw);
-            originalDarks.Add(name, mOrig.dark);
-
-
-            if (PlatformUtil.transformerLoaded)
-            {
-                double[] smoothed = PlatformUtil.ProcessBackground(m.wavenumbers, m.processed, "", 14, 200);
-                double[] wavenumbers = Enumerable.Range(400, smoothed.Length).Select(x => (double)x).ToArray();
-                Measurement updated = new Measurement();
-                updated.wavenumbers = wavenumbers;
-                updated.raw = smoothed;
-                library.Add(name, updated);
-
-#if USE_DECON
-                Deconvolution.Spectrum upSpec = new Deconvolution.Spectrum(new List<double>(wavenumbers), new List<double>(smoothed));
-                deconvolutionLibrary.library.Add(name, upSpec);
-#endif
-            }
-            else
-            {
-                Measurement updated = wavecal.crossMapIntensityWavenumber(otherCal, m);
-                double airPLSLambda = 10000;
-                int airPLSMaxIter = 100;
-                double[] array = AirPLS.smooth(updated.processed, airPLSLambda, airPLSMaxIter, 0.001, verbose: false, (int)roiStart, (int)roiEnd);
-                double[] shortened = new double[updated.processed.Length];
-                Array.Copy(array, 0, shortened, roiStart, array.Length);
-                updated.raw = shortened;
-                updated.dark = null;
-
-                library.Add(name, updated);
-            }
-
-            logger.info("finish loading library file from {0}", file.AbsolutePath);
-        }
+        
         public override async Task<Tuple<string,double>> findMatch(Measurement spectrum)
         {
             logger.debug("Library.findMatch: trying to match spectrum");
